API test naming (WIP)
