+++ cta_refinement/static/files/jsonSession/sessions.json
@@ -1,5 +1,1 @@
-<<<<<<< HEAD
-[{"session": "Cta User = {\r\nInit q0;\r\nq0 UM!card q1;\r\nq1 MU?pinrequest({x}) q2;\r\nq2 UM!pin(x <= 30) q3;\r\nq3 MU?menu q4;\r\n};\r\n\r\nCta User1 = {\r\nInit q0;\r\nq0 UM!card(x == 0) q1;\r\nq1 MU?pinrequest({x}) q2;\r\nq2 UM!pin(x == 0) q3;\r\nq3 MU?menu q4;\r\n}; \r\n\r\nUser1 refines? User;\r\n\r\n", "id": 690230366}, {"session": "Cta User = {\r\nInit q0;\r\nq0 UM!card q1;\r\nq1 MU?pinrequest({x}) q2;\r\nq2 UM!pin(x <= 30) q3;\r\nq3 MU?menu q4;\r\n};\r\n\r\nCta User1 = {\r\nInit q0;\r\nq0 UM!card(x == 0) q1;\r\nq1 MU?pinrequest({x}) q2;\r\nq2 UM!pin(x == 0) q3;\r\nq3 MU?menu q4;\r\n}; \r\n\r\nUser1 refines? User;\r\n\r\n", "id": 3519252470}, {"session": "Cta User = {\r\nInit q0;\r\nq0 UM!card q1;\r\nq1 MU?pinrequest({x}) q2;\r\nq2 UM!pin(x <= 30) q3;\r\nq3 MU?menu q4;\r\n};\r\n\r\nCta User1 = {\r\nInit q0;\r\nq0 UM!card(x == 0) q1;\r\nq1 MU?pinrequest({x}) q2;\r\nq2 UM!pin(x == 0) q3;\r\nq3 MU?menu q4;\r\n}; \r\n\r\nUser1 refines? User;\r\n\r\n", "id": 2585685421}, {"session": "Cta A = {\r\nInit start;\r\nstart CS?login logged;\r\nlogged CS?preApproval preApprovalApplication;\r\nlogged CS?selectVehicle vehicleSelection;\r\npreApprovalApplication SC!reject applicationRejected;\r\npreApprovalApplication SC!approved({x}) creditApproved;\r\ncreditApproved CS?selectVehicle(x < 30) vehicleSelection;\r\ncreditApproved SC!expired(x == 30) creditExpired;\r\nvehicleSelection CS?estimatePayment({x}) paymentEstimation;\r\nvehicleSelection CS?cancel cancelled;\r\npaymentEstimation CS?reapplication(x < 30) vehicleSelection;\r\npaymentEstimation CS?fullCredit(x <= 24) creditApplication;\r\npaymentEstimation SC!expired(x == 30) cancelled;\r\ncreditApplication SC!reject applicationRejected;\r\ncreditApplication SC!accept creditAccepted;\r\n};\r\n\r\nShow(A);\r\n\r\nCta A1 = {\r\nInit start;\r\nstart CS?login logged;\r\nlogged CS?preApproval preApprovalApplication;\r\nlogged CS?selectVehicle vehicleSelection;\r\npreApprovalApplication SC!reject (x == 0) applicationRejected;\r\npreApprovalApplication SC!approved(x == 0,{x}) creditApproved;\r\ncreditApproved CS?selectVehicle(x < 30 & x >= 29) vehicleSelection;\r\ncreditApproved SC!expired(x == 30) creditExpired;\r\nvehicleSelection CS?estimatePayment({x}) paymentEstimation;\r\nvehicleSelection CS?cancel cancelled;\r\npaymentEstimation CS?reapplication(x < 30 & x >= 29) vehicleSelection;\r\npaymentEstimation CS?fullCredit(x <= 24 & x >= 23) creditApplication;\r\npaymentEstimation SC!expired(x == 30) cancelled;\r\ncreditApplication SC!reject(x == 0) applicationRejected;\r\ncreditApplication SC!accept(x == 0) creditAccepted;\r\n};\r\n\r\nCta A2 = {\r\nInit start;\r\nstart CS?login logged;\r\nlogged CS?preApproval preApprovalApplication;\r\nlogged CS?selectVehicle vehicleSelection;\r\npreApprovalApplication SC!reject applicationRejected;\r\npreApprovalApplication SC!approved({x}) creditApproved;\r\ncreditApproved CS?selectVehicle(x < 30 & x >= 29) vehicleSelection;\r\ncreditApproved SC!expired(x == 30) creditExpired;\r\nvehicleSelection CS?estimatePayment({x}) paymentEstimation;\r\nvehicleSelection CS?cancel cancelled;\r\npaymentEstimation CS?reapplication(x < 30 & x >= 29) vehicleSelection;\r\npaymentEstimation CS?fullCredit(x <= 24 & x >= 23) creditApplication;\r\npaymentEstimation SC!expired(x == 30) cancelled;\r\ncreditApplication SC!reject applicationRejected;\r\ncreditApplication SC!accept creditAccepted;\r\n};\r\n\r\nA1 refines? A;\r\n\r\nA2 refines? A;\r\n", "id": 2130155073}]
-=======
-[{"session": "Cta User = {\r\nInit q0;\r\nq0 UM!card q1;\r\nq1 MU?pinrequest({x}) q2;\r\nq2 UM!pin(x <= 30) q3;\r\nq3 MU?menu q4;\r\n};\r\n\r\nCta User1 = {\r\nInit q0;\r\nq0 UM!card(x == 0) q1;\r\nq1 MU?pinrequest({x}) q2;\r\nq2 UM!pin(x == 0) q3;\r\nq3 MU?menu q4;\r\n}; \r\n\r\nUser1 refines? User;\r\n\r\n", "id": 690230366}, {"session": "Cta User = {\r\nInit q0;\r\nq0 UM!card q1;\r\nq1 MU?pinrequest({x}) q2;\r\nq2 UM!pin(x <= 30) q3;\r\nq3 MU?menu q4;\r\n};\r\n\r\nCta User1 = {\r\nInit q0;\r\nq0 UM!card(x == 0) q1;\r\nq1 MU?pinrequest({x}) q2;\r\nq2 UM!pin(x == 0) q3;\r\nq3 MU?menu q4;\r\n}; \r\n\r\nUser1 refines? User;\r\n\r\n", "id": 3519252470}, {"session": "Cta User = {\r\nInit q0;\r\nq0 UM!card q1;\r\nq1 MU?pinrequest({x}) q2;\r\nq2 UM!pin(x <= 30) q3;\r\nq3 MU?menu q4;\r\n};\r\n\r\nCta User1 = {\r\nInit q0;\r\nq0 UM!card(x == 0) q1;\r\nq1 MU?pinrequest({x}) q2;\r\nq2 UM!pin(x == 0) q3;\r\nq3 MU?menu q4;\r\n}; \r\n\r\nUser1 refines? User;\r\n\r\n", "id": 2585685421}, {"session": "Cta User = {\r\nInit q0;\r\nq0 UM!card q1;\r\nq1 MU?pinrequest({x}) q2;\r\nq2 UM!pin(x <= 30) q3;\r\nq3 MU?menu q4;\r\n};\r\n\r\nCta User1 = {\r\nInit q0;\r\nq0 UM!card(x == 0) q1;\r\nq1 MU?pinrequest({x}) q2;\r\nq2 UM!pin(x == 0) q3;\r\nq3 MU?menu q4;\r\n}; \r\n\r\nUser1 refines? User;\r\n\r\n", "id": 439722809}]
->>>>>>> 75c0b958ac2bffe87fc3c639afecf23f01517d7e
+[{"session": "Cta User = {\r\nInit q0;\r\nq0 UM!card q1;\r\nq1 MU?pinrequest({x}) q2;\r\nq2 UM!pin(x <= 30) q3;\r\nq3 MU?menu q4;\r\n};\r\n\r\nCta User1 = {\r\nInit q0;\r\nq0 UM!card(x == 0) q1;\r\nq1 MU?pinrequest({x}) q2;\r\nq2 UM!pin(x == 0) q3;\r\nq3 MU?menu q4;\r\n}; \r\n\r\nUser1 refines? User;\r\n\r\n", "id": 1971674163}]
