+++ cta_refinement/tests/test_test.py
@@ -1,0 +1,3 @@
+def test_one():
+   assert 3 == 5
+
